Simplify; don't try to be clever
Instead of trying to delay creation of Luxafor instances etc., we now just create a new device and set the current colors on each update.

Some of the reason for the early returns etc. before was because we'd also blink the lights when the state changed. Since we no longer do that, the code can be simplified a lot.